Began reworking UI Placements
Root of cause: Window and many widgets are set in a fix dimension, setGeometry(). It needs to use
a more dynamic sizing framework to allow scrolling in the playlist tab.

Solution (for now): Destroy and Rebuild! Having the same design as before but with layout widgets,
widgets that are meant to do the resizing for you at the cost of fixed sizes.